requirements.txt
atproto
python-dotenv
requests
asyncio
aiohttp
ffmpeg-python
